[Refactor] Ingest Layer 패키지 조정 및 주석, 로그 조정
