+++ src/main/java/com/logmate/streaming/ingress/GzipDecodingWebFilter.java
@@ -1,4 +1,4 @@
-package com.logmate.streaming.common.webfilter;
+package com.logmate.streaming.ingress;
 
 import java.io.ByteArrayInputStream;
 import java.io.IOException;
@@ -23,42 +23,60 @@ import org.springframework.web.server.WebFilterChain;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 
+/**
+ * GzipDecodingWebFilter
+ * <p>
+ * 요청 본문이 Gzip 압축되어 있을 경우 자동으로 해제하여 downstream handler가 일반 JSON 요청처럼 처리할 수 있도록 변환하는 WebFilter.
+ * <p>
+ * Reactive Stream 환경에서는 요청 body가 여러 DataBuffer로 분리되어 들어올 수 있기 때문에 DataBufferUtils.join()으로 전체를 모은 뒤
+ * 압축 해제를 수행한다.
+ * <p>
+ * 압축 해제 후 Content-Encoding 헤더를 제거하고 Content-Type을 application/json으로 지정한다.
+ */
 @Slf4j
 @Component
 @Order(Ordered.HIGHEST_PRECEDENCE + 10)
 @RequiredArgsConstructor
 public class GzipDecodingWebFilter implements WebFilter {
+
   private static final List<String> GZIP_CT = List.of(
       "application/gzip",
       "application/x-gzip"
   );
 
-  @Value("${web.gzip.max-bytes:10485760}") // 기본 10MB
+  /**
+   * 최대 허용 바이트 (기본 10MB)
+   */
+  @Value("${web.gzip.max-bytes:10485760}")
   private int maxBytes;
 
   @Override
   public Mono<Void> filter(ServerWebExchange exchange, WebFilterChain chain) {
-    log.info("[GzipDecodingWebFilter] filter called");
     HttpHeaders headers = exchange.getRequest().getHeaders();
     String contentEncoding = headers.getFirst(HttpHeaders.CONTENT_ENCODING);
-    String contentType = headers.getContentType() != null ? headers.getContentType().toString() : null;
+    String contentType =
+        headers.getContentType() != null ? headers.getContentType().toString() : null;
 
     boolean isGzip =
         (contentEncoding != null && "gzip".equalsIgnoreCase(contentEncoding))
-            || (contentType != null && GZIP_CT.stream().anyMatch(ct -> ct.equalsIgnoreCase(contentType)));
+            || (contentType != null && GZIP_CT.stream()
+            .anyMatch(ct -> ct.equalsIgnoreCase(contentType)));
 
     if (!isGzip) {
-      // gzip 아니면 그대로 진행
+      // gzip이 아닌 경우 그대로 통과
       return chain.filter(exchange);
     }
 
+    log.info("[GzipDecodingWebFilter] Detected gzip request (path={}, encoding={})",
+        exchange.getRequest().getURI(), contentEncoding);
     // 바디를 한 번 모아서(사이즈 체크) gunzip → JSON으로 변경
     return DataBufferUtils.join(exchange.getRequest().getBody())
         .flatMap(joined -> {
           int readable = joined.readableByteCount();
           if (readable > maxBytes) {
             DataBufferUtils.release(joined);
-            return Mono.error(new IllegalArgumentException("Compressed body too large"));
+            return Mono.error(new IllegalArgumentException(
+                "[GzipDecodingWebFilter] Compressed body too large (" + readable + " bytes)"));
           }
           byte[] compressed = new byte[readable];
           joined.read(compressed);
@@ -68,11 +86,13 @@ public class GzipDecodingWebFilter implements WebFilter {
           try {
             decompressed = gunzip(compressed, maxBytes);
           } catch (IOException e) {
-            return Mono.error(new IllegalArgumentException("Failed to decompress gzip body", e));
+            return Mono.error(new IllegalArgumentException(
+                "[GzipDecodingWebFilter] Failed to decompress gzip body", e));
           }
 
           if (decompressed.length > maxBytes) {
-            return Mono.error(new IllegalArgumentException("Decompressed body too large"));
+            return Mono.error(new IllegalArgumentException(
+                "[GzipDecodingWebFilter] Decompressed body too large (\" + decompressed.length + \" bytes)\")"));
           }
 
           DataBufferFactory factory = exchange.getResponse().bufferFactory();
@@ -99,12 +119,13 @@ public class GzipDecodingWebFilter implements WebFilter {
             }
           };
 
+          log.info("[GzipDecodingWebFilter] Decompressed {} → {} bytes (path={})",
+              readable, decompressed.length, exchange.getRequest().getURI());
           return chain.filter(exchange.mutate().request(decorated).build());
         });
   }
 
   private byte[] gunzip(byte[] compressed, int limit) throws IOException {
-    log.info("[GzipDecodingWebFilter] gunzip called");
     try (GZIPInputStream gis = new GZIPInputStream(new ByteArrayInputStream(compressed))) {
       // 안전한 확장: limit까지 읽기
       byte[] buf = new byte[Math.min(8192, limit)];
@@ -113,7 +134,7 @@ public class GzipDecodingWebFilter implements WebFilter {
       int r;
       while ((r = gis.read(buf)) != -1) {
         if (total + r > limit) {
-          throw new IOException("Decompressed size exceeds limit");
+          throw new IOException("[GzipDecodingWebFilter] Decompressed size exceeds limit");
         }
         byte[] newOut = new byte[total + r];
         System.arraycopy(out, 0, newOut, 0, total);
